requirements.txt
discord.py
asyncio
black
isort
chess
humanfriendly
flask
